refactor: get project items to data
