+++ src/app/resume-page/data/get-project-items.ts
@@ -1,7 +1,7 @@
 import { inject, InjectionToken } from '@angular/core'
-import { Project } from '../data/project'
-import { ADAPT_JSON_RESUME_PROJECT } from '../data/adapt-json-resume-project'
-import { JSON_RESUME_PROJECTS } from './json-resume-projects'
+import { Project } from './project'
+import { ADAPT_JSON_RESUME_PROJECT } from './adapt-json-resume-project'
+import { JSON_RESUME_PROJECTS } from '../projects-section/json-resume-projects'
 
 export type GetProjectItems = () => readonly Project[]
