+++ src/app/resume-page/plain-resume/plain-resume.component.ts
@@ -2,7 +2,7 @@ import { ChangeDetectionStrategy, Component, inject } from '@angular/core'
 import { LinkComponent } from '../link/link.component'
 import { DateRangeComponent } from '../date-range/date-range.component'
 import { EDUCATION_SERVICE } from '../data/education-service'
-import { GET_PROJECT_ITEMS } from '../projects-section/get-project-items'
+import { GET_PROJECT_ITEMS } from '../data/get-project-items'
 import { ContentPageComponent } from '../../content-page/content-page.component'
 import { MdLinksPipe } from '../md-links.pipe'
 import { MaterialSymbolDirective } from '@/common/material-symbol.directive'
