+++ src/app/resume-page/projects-section/projects-section.component.spec.ts
@@ -3,7 +3,7 @@ import { ComponentFixture } from '@angular/core/testing'
 import { ProjectsSectionComponent } from './projects-section.component'
 import { MockComponents, MockProvider } from 'ng-mocks'
 import { SectionTitleComponent } from '../section-title/section-title.component'
-import { GET_PROJECT_ITEMS, GetProjectItems } from './get-project-items'
+import { GET_PROJECT_ITEMS, GetProjectItems } from '../data/get-project-items'
 import { ProjectItemComponent } from './project-item/project-item.component'
 import { componentTestSetup } from '@/test/helpers/component-test-setup'
 import { makeProject } from '../data/__tests__/make-project'
